Conflicts:
	desapp-groupf-backend/src/main/java/model/User.java
	desapp-groupf-backend/src/main/java/persistence/UserRepository.java
	desapp-groupf-backend/src/main/java/webservice/UserRest.java
	desapp-groupf-backend/src/main/resources/import.sql

+++ desapp-groupf-backend/src/main/java/service/GenericService.java
@@ -49,7 +49,7 @@ public class GenericService<T> implements Serializable {
     }
     
     @Transactional
-    public T getById(long id){
+    public T getById(Integer id){
     	return this.getRepository().findById(id);
     }
